moved functions to utils directory
